Fix reviews : - Catch exceptions from Symfony/mailer - Fixed somes use - Updated the checking encryption in TransportFactory - Renamed MailerTest into SymfonyMailerTest
