+++ main/core/Resources/config/services/library.yml
@@ -94,6 +94,7 @@ services:
         tags: ['claroline.mailing']
         arguments:
             - '@mailer'
+            - '@logger'
 
     Claroline\CoreBundle\Library\Mailing\Client\Postal:
         tags: ['claroline.mailing']
