+++ main/core/Resources/config/suggested/mailer_dev.yml
@@ -1,0 +1,2 @@
+imports:
+    - { resource: mailer_prod.yml }
