+++ main/core/Resources/config/suggested/mailer_prod.yml
@@ -1,0 +1,2 @@
+framework:
+    mailer: ~
