--- main/core/Resources/config/suggested/swiftmailer_dev.yml
@@ -1,2 +1,0 @@
-imports:
-    - { resource: swiftmailer_prod.yml }
