--- main/core/Resources/config/suggested/swiftmailer_prod.yml
@@ -1,8 +1,0 @@
-swiftmailer:
-    transport:  "%mailer_transport%"
-    encryption: "%mailer_encryption%"
-    auth_mode:  "%mailer_auth_mode%"
-    host:       "%mailer_host%"
-    port:       "%mailer_port%"
-    username:   "%mailer_user%"
-    password:   "%mailer_password%"
