--- main/core/Resources/config/suggested/swiftmailer_test.yml
@@ -1,5 +1,0 @@
-imports:
-    - { resource: swiftmailer_dev.yml }
-
-swiftmailer:
-    disable_delivery: true
